Try again: cast terabyte & petabyte to (uint64_t), that's how we're using them.

+++ src/include/misc.h
@@ -12,8 +12,8 @@
 #define	WT_KILOBYTE	(1024)
 #define	WT_MEGABYTE	(1048576)
 #define	WT_GIGABYTE	(1073741824)
-#define	WT_TERABYTE	(1099511627776ULL)
-#define	WT_PETABYTE	(1125899906842624ULL)
+#define	WT_TERABYTE	((uint64_t)1099511627776)
+#define	WT_PETABYTE	((uint64_t)1125899906842624)
 
 /*
  * Sizes that cannot be larger than 2**32 are stored in uint32_t fields in
